Refactor: Decreased Handler Function Call Length
I have decreases the long import statements for the handlers by not
representing the functions as functions but external parts of my code
and dynamically importing the side effects from them.
There is also a start.ps1 script for windows users that replicates the
start.sh script I created. README is also updated.